cleanup for arm64

+++ lib/cocos2d-x/cocos2dx/platform/ios/EAGLView.mm
@@ -399,7 +399,7 @@ static EAGLView *view = 0;
         return;
     }
     
-    size_t ids[IOS_MAX_TOUCHES_COUNT] = {0};
+    int ids[IOS_MAX_TOUCHES_COUNT] = {0};
     float xs[IOS_MAX_TOUCHES_COUNT] = {0.0f};
     float ys[IOS_MAX_TOUCHES_COUNT] = {0.0f};
     
@@ -419,7 +419,7 @@ static EAGLView *view = 0;
     {
         return;
     }
-    size_t ids[IOS_MAX_TOUCHES_COUNT] = {0};
+    int ids[IOS_MAX_TOUCHES_COUNT] = {0};
     float xs[IOS_MAX_TOUCHES_COUNT] = {0.0f};
     float ys[IOS_MAX_TOUCHES_COUNT] = {0.0f};
     
@@ -440,7 +440,7 @@ static EAGLView *view = 0;
         return;
     }
     
-    size_t ids[IOS_MAX_TOUCHES_COUNT] = {0};
+    int ids[IOS_MAX_TOUCHES_COUNT] = {0};
     float xs[IOS_MAX_TOUCHES_COUNT] = {0.0f};
     float ys[IOS_MAX_TOUCHES_COUNT] = {0.0f};
     
@@ -461,7 +461,7 @@ static EAGLView *view = 0;
         return;
     }
     
-    size_t ids[IOS_MAX_TOUCHES_COUNT] = {0};
+    int ids[IOS_MAX_TOUCHES_COUNT] = {0};
     float xs[IOS_MAX_TOUCHES_COUNT] = {0.0f};
     float ys[IOS_MAX_TOUCHES_COUNT] = {0.0f};
